Correct ECONNREFUSED error
Previous commits experienced an ECONNREFUSED error when the client tries
to connect to the server.

Investigating this issue, I discovered that this issue is known (to the
SAFE-template team); however, it seems as though this work did not
migrate to the SAFE-dojo project.

This commit changes the server to listen on **all** interfaces (today,
on both IPv4 and IPv6) for requests.